+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - HR: changed fromDate and toDate to fromDateT and toDateT in insertLeave method of HumanResourceMobileController
 - MOVE LINE: fix amount paid display in move line form.
 - Ebics User : resolve error getting on export & Modify import config and export template to include BankOrderList and BankStatementList of EbicsPartner.
+- Sale order/quotation: fix tab title when generating a quotation from an opportunity.
 
 ## [5.1.13] - 2019-11-29
 ## Improvements